+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2025-03-11
+
+### Added
+
+- Add `vscode-marketplace-client` dependency to check for extension updates and display a notification
+
+### Changed
+
+- Update the `extension.ts` file to use the new `vscode-marketplace-client` dependency
+- Update Localization strings for the extension
+
 ## [2.3.0] - 2025-01-10
 
 ### Added
@@ -161,7 +172,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v2.2.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v2.4.0...HEAD
+[2.4.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v2.3.0...v2.4.0
+[2.3.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.8.1...v2.0.0